+++ src/jp/ac/oit/igakilab/tasks/test/TestTrelloActions.java
@@ -12,8 +12,8 @@ import java.util.Set;
 import org.json.simple.JSONArray;
 import org.json.simple.JSONObject;
 
+import jp.ac.oit.igakilab.tasks.trello.api.SimpleJsonResponseTextParser;
 import jp.ac.oit.igakilab.tasks.trello.api.TrelloApi;
-import jp.ac.oit.igakilab.tasks.trello.api.TrelloApi.SimpleJsonResponseTextParser;
 
 public class TestTrelloActions {
 	public static String TRELLO_API_KEY = "67ad72d3feb45f7a0a0b3c8e1467ac0b";
